+++ app/src/main/java/org/simple/clinic/recentpatient/AllRecentPatientsModel.kt
@@ -2,11 +2,25 @@ package org.simple.clinic.recentpatient
 
 import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
+import org.simple.clinic.patient.RecentPatient
 
 @Parcelize
-class AllRecentPatientsModel: Parcelable {
+data class AllRecentPatientsModel(
+    val recentPatients: List<RecentPatient>?
+): Parcelable {
 
   companion object {
-    fun create(): AllRecentPatientsModel = AllRecentPatientsModel()
+    fun create(): AllRecentPatientsModel {
+      return AllRecentPatientsModel(
+          recentPatients = null
+      )
+    }
   }
+
+  fun recentPatientsLoaded(recentPatients: List<RecentPatient>): AllRecentPatientsModel {
+    return copy(recentPatients = recentPatients)
+  }
+
+  val hasLoadedRecentPatients: Boolean
+    get() = recentPatients != null
 }