+++ app/src/main/java/org/simple/clinic/recentpatient/RecentPatientsScreenEvents.kt
@@ -1,8 +1,0 @@
-package org.simple.clinic.recentpatient
-
-import org.simple.clinic.widgets.UiEvent
-import java.util.UUID
-
-data class RecentPatientItemClicked(val patientUuid: UUID) : UiEvent {
-  override val analyticsName = "Recent Patients: Item clicked"
-}