+++ app/src/main/java/org/simple/clinic/recentpatientsview/LatestRecentPatientsUiRenderer.kt
@@ -3,11 +3,7 @@ package org.simple.clinic.recentpatientsview
 import com.squareup.inject.assisted.Assisted
 import com.squareup.inject.assisted.AssistedInject
 import org.simple.clinic.mobius.ViewRenderer
-import org.simple.clinic.patient.DateOfBirth
-import org.simple.clinic.patient.RecentPatient
 import org.simple.clinic.util.UserClock
-import org.simple.clinic.util.toLocalDateAtZone
-import java.time.LocalDate
 import java.time.format.DateTimeFormatter
 import javax.inject.Named
 
@@ -31,10 +27,8 @@ class LatestRecentPatientsUiRenderer @AssistedInject constructor(
   }
 
   private fun renderRecentPatients(model: LatestRecentPatientsModel) {
-    val today = LocalDate.now(userClock)
-
     val recentPatientItems = addSeeAllIfListTooLong(
-        recentPatients = model.recentPatients!!.map { recentPatientItem(it, today) },
+        recentPatients = RecentPatientItemType.create(model.recentPatients!!, userClock, dateFormatter),
         recentPatientLimit = numberOfPatientsToShow
     )
 
@@ -49,28 +43,8 @@ class LatestRecentPatientsUiRenderer @AssistedInject constructor(
     }
   }
 
-  private fun recentPatientItem(recentPatient: RecentPatient, today: LocalDate): RecentPatientItem {
-    val patientRegisteredOnDate = recentPatient.patientRecordedAt.toLocalDateAtZone(userClock.zone)
-    val isNewRegistration = today == patientRegisteredOnDate
-
-    return RecentPatientItem(
-        uuid = recentPatient.uuid,
-        name = recentPatient.fullName,
-        age = age(recentPatient),
-        gender = recentPatient.gender,
-        updatedAt = recentPatient.updatedAt,
-        dateFormatter = dateFormatter,
-        clock = userClock,
-        isNewRegistration = isNewRegistration
-    )
-  }
-
-  private fun age(recentPatient: RecentPatient): Int {
-    return DateOfBirth.fromRecentPatient(recentPatient, userClock).estimateAge(userClock)
-  }
-
   private fun addSeeAllIfListTooLong(
-      recentPatients: List<RecentPatientItem>,
+      recentPatients: List<RecentPatientItemType>,
       recentPatientLimit: Int
   ) = if (recentPatients.size > recentPatientLimit) {
     recentPatients.take(recentPatientLimit) + SeeAllItem