+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 - Migrated `RecentPatientsView` to Mobius
 - Migrated `RegistrationLoadingScreen` to Mobius
 - Change selection of facility throughout the codebase to use `ItemAdapter`
-- [In Progress: 21st Jul 2020] Migrate `RecentPatientsScreen` to Mobius
+- [In Progress: 23rd Jul 2020] Migrate `RecentPatientsScreen` to Mobius
 - [In Progress: 16th Jul 2020] Migrate `LoginPinScreen` to Mobius
 
 ### Changes